[Improve] Migrate Admin swagger from springfox to springdoc
