+++ shenyu-admin/src/test/java/org/apache/shenyu/admin/service/manager/impl/SwaggerDocParserTest.java
@@ -34,9 +34,9 @@ public class SwaggerDocParserTest {
     public static final String DOC_INFO_JSON = "{\n"
         + "    \"swagger\":\"2.0\",\n"
         + "    \"info\":{\n"
-        + "        \"description\":\"shenyu-examples-http-swagger2 API\",\n"
+        + "        \"description\":\"shenyu-examples-http-swagger3 API\",\n"
         + "        \"version\":\"2.3.0\",\n"
-        + "        \"title\":\"shenyu-examples-http-swagger2 API\",\n"
+        + "        \"title\":\"shenyu-examples-http-swagger3 API\",\n"
         + "        \"contact\":{\n"
         + "            \"name\":\"ShenYu\",\n"
         + "            \"url\":\"https://github.com/apache/shenyu\",\n"
