+++ shenyu-examples/pom.xml
@@ -38,13 +38,13 @@
         <versions-maven-plugin.version>2.5</versions-maven-plugin.version>
         <docker-maven-plugin.version>0.40.1</docker-maven-plugin.version>
         <maven-checkstyle-plugin.version>3.1.0</maven-checkstyle-plugin.version>
-        <swagger.version>2.9.2</swagger.version>
+        <swagger.version>2.6.0</swagger.version>
     </properties>
 
     <modules>
         <module>shenyu-examples-common</module>
         <module>shenyu-examples-http</module>
-        <module>shenyu-examples-http-swagger2</module>
+        <module>shenyu-examples-http-swagger3</module>
         <module>shenyu-examples-https</module>
         <module>shenyu-examples-dubbo</module>
         <module>shenyu-examples-springcloud</module>
@@ -63,14 +63,8 @@
     <dependencyManagement>
         <dependencies>
             <dependency>
-                <groupId>io.springfox</groupId>
-                <artifactId>springfox-swagger2</artifactId>
-                <version>${swagger.version}</version>
-            </dependency>
-
-            <dependency>
-                <groupId>io.springfox</groupId>
-                <artifactId>springfox-swagger-ui</artifactId>
+                <groupId>org.springdoc</groupId>
+                <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
                 <version>${swagger.version}</version>
             </dependency>
         </dependencies>
