--- shenyu-examples/shenyu-examples-http-swagger2/k8s/shenyu-examples-http-service.yml
@@ -1,62 +1,0 @@
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-apiVersion: v1
-kind: Service
-metadata:
-  name: shenyu-examples-http-swagger2-service
-  labels:
-    app: shenyu-examples-http-swagger2
-    all: shenyu-examples-http-swagger2
-spec:
-  selector:
-    app: shenyu-examples-http-swagger2
-    all: shenyu-examples-http-swagger2
-  type: NodePort
-  ports:
-  - name: "8190"
-    port: 8190
-    targetPort: 8190
-    nodePort: 31189
-status:
-  loadBalancer: {}
-
----
-apiVersion: v1
-kind: Service
-metadata:
-  name: shenyu-zk
-  labels:
-    app: shenyu-zk
-    all: shenyu-examples-http-swagger2
-spec:
-  type: NodePort
-  selector:
-    app: shenyu-zk
-    all: shenyu-examples-http-swagger2
-  ports:
-    - name: "client"
-      port: 2181
-      targetPort: 2181
-    - name: "server"
-      port: 2888
-      targetPort: 2888
-    - name: "election"
-      port: 3888
-      targetPort: 3888
-    - name: "website"
-      port: 8080
-      targetPort: 8080
