--- shenyu-examples/shenyu-examples-http-swagger2/src/main/java/org/apache/shenyu/examples/http/controller/UploadController.java
@@ -1,65 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.shenyu.examples.http.controller;
-
-import io.swagger.annotations.Api;
-import io.swagger.annotations.ApiOperation;
-import java.util.stream.Collectors;
-import org.apache.shenyu.client.springmvc.annotation.ShenyuSpringMvcClient;
-import org.springframework.http.MediaType;
-import org.springframework.http.codec.multipart.FilePart;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestPart;
-import org.springframework.web.bind.annotation.RestController;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-/**
- * UploadController.
- */
-@RestController
-@RequestMapping("/upload")
-@ShenyuSpringMvcClient("/upload/**")
-@Api(tags = "upload api")
-public class UploadController {
-
-    /**
-     * webFlux uploadFile.
-     *
-     * @param file the file you upload
-     * @return response
-     */
-    @PostMapping(value = "/webFluxSingle", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE, MediaType.TEXT_PLAIN_VALUE})
-    @ApiOperation(value = "webFluxSingle", notes = "upload webFluxSingle.")
-    public Mono<String> webFluxSingle(@RequestPart("file") final FilePart file) {
-        return Mono.just(file.filename());
-    }
-
-    /**
-     * webFlux uploadFiles.
-     *
-     * @param files the file you upload
-     * @return response
-     */
-    @PostMapping(value = "/webFluxFiles", consumes = {MediaType.MULTIPART_FORM_DATA_VALUE, MediaType.TEXT_PLAIN_VALUE})
-    @ApiOperation(value = "webFluxFiles", notes = "upload webFluxFiles.")
-    public Mono<String> webFluxFiles(@RequestPart(value = "files", required = false) final Flux<FilePart> files) {
-        return files.map(FilePart::filename).collect(Collectors.joining(","));
-    }
-}
