+++ shenyu-examples/shenyu-examples-http-swagger3/k8s/script/healthcheck.sh
