+++ shenyu-examples/shenyu-examples-http-swagger3/k8s/shenyu-deployment.yml
@@ -1,0 +1,82 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: shenyu-admin
+  labels:
+    app: shenyu-admin
+    all: shenyu-examples-http-swagger3
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: shenyu-admin
+      all: shenyu-examples-http-swagger3
+  template:
+    metadata:
+      labels:
+        app: shenyu-admin
+        all: shenyu-examples-http-swagger3
+    spec:
+      containers:
+        - name: shenyu-admin
+          image: apache/shenyu-admin:latest
+          resources: {}
+          env:
+            - name: SPRING_PROFILES_ACTIVE
+              value: h2
+            - name: shenyu.database.init_script
+              value: sql-script/h2/schema.sql
+          ports:
+            - containerPort: 9095
+          imagePullPolicy: IfNotPresent
+      restartPolicy: Always
+status: {}
+
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: shenyu-bootstrap
+  labels:
+    app: shenyu-bootstrap
+    all: shenyu-examples-http-swagger3
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: shenyu-bootstrap
+      all: shenyu-examples-http-swagger3
+  template:
+    metadata:
+      labels:
+        app: shenyu-bootstrap
+        all: shenyu-examples-http-swagger3
+    spec:
+      containers:
+        - name: shenyu-bootstrap
+          image: apache/shenyu-bootstrap:latest
+          resources: {}
+          env:
+            - name: shenyu.sync.websocket.urls
+              value: ws://shenyu-admin:9095/websocket
+          ports:
+            - containerPort: 9195
+          imagePullPolicy: IfNotPresent
+      restartPolicy: Always
+status: {}
