+++ shenyu-examples/shenyu-examples-http-swagger3/k8s/shenyu-examples-http-deployment.yml
@@ -1,0 +1,91 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: shenyu-examples-http-swagger3-deployment
+  labels:
+    app: shenyu-examples-http-swagger3
+    all: shenyu-examples-http-swagger3
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: shenyu-examples-http-swagger3
+      all: shenyu-examples-http-swagger3
+  strategy: {}
+  template:
+    metadata:
+      labels:
+        app: shenyu-examples-http-swagger3
+        all: shenyu-examples-http-swagger3
+    spec:
+      containers:
+        - image: shenyu-examples-http-swagger3
+          name: shenyu-examples-http-swagger3
+          livenessProbe:
+            exec:
+              command:
+                - wget -q -O - http://localhost:8190/actuator/health | grep UP || exit 1
+            initialDelaySeconds: 10
+            failureThreshold: 3
+            timeoutSeconds: 2
+          env:
+            - name: shenyu.register.serverLists
+              value: http://shenyu-admin:9095
+          ports:
+            - containerPort: 8190
+          imagePullPolicy: IfNotPresent
+      restartPolicy: Always
+status: {}
+
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  labels:
+    app: shenyu-zk
+    all: shenyu-examples-http-swagger3
+  name: shenyu-zk
+spec:
+  replicas: 1
+  selector:
+    matchLabels:
+      app: shenyu-zk
+      all: shenyu-examples-http-swagger3
+  strategy: {}
+  template:
+    metadata:
+      labels:
+        app: shenyu-zk
+        all: shenyu-examples-http-swagger3
+    spec:
+      containers:
+        - image: zookeeper:3.5
+          name: shenyu-zk
+          resources: {}
+          ports:
+            - containerPort: 2181
+              name: client
+            - containerPort: 2888
+              name: server
+            - containerPort: 3888
+              name: leader-election
+            - containerPort: 8080
+              name: website
+      restartPolicy: Always
+status: {}
