+++ shenyu-examples/shenyu-examples-http-swagger3/k8s/shenyu-service.yml
@@ -1,0 +1,52 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+apiVersion: v1
+kind: Service
+metadata:
+  name: shenyu-admin
+  labels:
+    app: shenyu-admin
+    all: shenyu-examples-http-swagger3
+spec:
+  type: NodePort
+  selector:
+    app: shenyu-admin
+    all: shenyu-examples-http-swagger3
+  ports:
+  - name: "9095"
+    port: 9095
+    targetPort: 9095
+    nodePort: 31095
+
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: shenyu-bootstrap
+  labels:
+    app: shenyu-bootstrap
+    all: shenyu-examples-http-swagger3
+spec:
+  type: NodePort
+  selector:
+    app: shenyu-bootstrap
+    all: shenyu-examples-http-swagger3
+  ports:
+  - name: "9195"
+    port: 9195
+    targetPort: 9195
+    nodePort: 31195
