+++ shenyu-examples/shenyu-examples-http-swagger3/pom.xml
@@ -25,7 +25,7 @@
         <version>2.7.0-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
-    <artifactId>shenyu-examples-http</artifactId>
+    <artifactId>shenyu-examples-http-swagger3</artifactId>
 
     <properties>
         <commons-lang3.version>3.12.0</commons-lang3.version>
@@ -38,11 +38,6 @@
             <groupId>org.apache.shenyu</groupId>
             <artifactId>shenyu-spring-boot-starter-client-springmvc</artifactId>
             <version>${project.version}</version>
-        </dependency>
-
-        <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter</artifactId>
         </dependency>
 
 <!--        <dependency>-->
@@ -67,7 +62,19 @@
 <!--            <groupId>org.apache.shenyu</groupId>-->
 <!--            <artifactId>shenyu-discovery-eureka</artifactId>-->
 <!--            <version>${project.version}</version>-->
+<!--            <exclusions>-->
+<!--                <exclusion>-->
+<!--                    <groupId>javax.servlet</groupId>-->
+<!--                    <artifactId>servlet-api</artifactId>-->
+<!--                </exclusion>-->
+<!--            </exclusions>-->
 <!--        </dependency>-->
+
+
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter</artifactId>
+        </dependency>
 
         <dependency>
             <groupId>org.springframework.boot</groupId>
@@ -97,16 +104,25 @@
             <version>2.2.6.RELEASE</version>
         </dependency>-->
 
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-web</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springdoc</groupId>
+            <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
-        <finalName>shenyu-examples-http</finalName>
+        <finalName>shenyu-examples-http-swagger3</finalName>
         <plugins>
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
                 <configuration>
-                    <mainClass>org.apache.shenyu.examples.http.ShenyuTestHttpApplication</mainClass>
+                    <mainClass>org.apache.shenyu.examples.http.ShenyuTestSwaggerApplication</mainClass>
                     <executable>true</executable>
                 </configuration>
             </plugin>
@@ -117,8 +133,8 @@
         <profile>
             <id>example</id>
             <properties>
-                <docker.buildArg.APP_NAME>shenyu-examples-http</docker.buildArg.APP_NAME>
-                <docker.image.tag.repo>shenyu-examples-http</docker.image.tag.repo>
+                <docker.buildArg.APP_NAME>shenyu-examples-http-swagger3</docker.buildArg.APP_NAME>
+                <docker.image.tag.repo>shenyu-examples-http-swagger3</docker.image.tag.repo>
                 <docker.image.tag.tagName>latest</docker.image.tag.tagName>
             </properties>
             <activation>
@@ -133,7 +149,7 @@
                         <configuration>
                             <images>
                                 <image>
-                                    <name>shenyu-examples-http</name>
+                                    <name>shenyu-examples-http-swagger3</name>
                                     <build>
                                         <contextDir>${project.basedir}</contextDir>
                                     </build>
