+++ shenyu-examples/shenyu-examples-http-swagger3/src/main/resources/application.yml
@@ -1,0 +1,73 @@
+# Licensed to the Apache Software Foundation (ASF) under one or more
+# contributor license agreements.  See the NOTICE file distributed with
+# this work for additional information regarding copyright ownership.
+# The ASF licenses this file to You under the Apache License, Version 2.0
+# (the "License"); you may not use this file except in compliance with
+# the License.  You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+server:
+  port: 8190
+  address: 0.0.0.0
+  tomcat:
+    max-http-form-post-size: 100MB
+  servlet:
+    context-path: /http-swagger3 # should be the same as shenyu.client.http.contextPath
+
+shenyu:
+  register:
+    registerType: http
+    serverLists: http://127.0.0.1:9095
+    props:
+      username: admin
+      password: 123456
+  client:
+      http:
+        props:
+          contextPath: /http-swagger3
+          appName: http-swagger3
+#          port: 8190
+#  discovery:
+#    enable: true
+#    protocol: http:// #https://
+#    type: zookeeper
+#    serverList: 127.0.0.1:2181
+#    registerPath: /shenyu/discovery/demo_swagger3_common
+#    props:
+#      baseSleepTimeMilliseconds: 1000
+#      maxRetries: 4
+#      maxSleepTimeMilliseconds: 5000
+#      connectionTimeoutMilliseconds: 60000
+#      sessionTimeoutMilliseconds: 8
+#    type: etcd
+#    serverList: http://127.0.0.1:2379
+#    registerPath: shenyu_discovery_demo_${shenyu.client.http.props.appName}
+#    type: eureka
+#    serverList: http://127.0.0.1:8761/eureka
+#    registerPath: shenyu_discovery_demo_http-swagger3
+
+spring:
+  servlet:
+    multipart:
+      max-file-size: 100MB
+      max-request-size: 100MB
+  mvc:
+    pathmatch:
+      matching-strategy: ant_path_matcher
+
+logging:
+  level:
+    root: info
+    org.springframework.boot: info
+    org.apache.ibatis: info
+    org.apache.shenyu.test.bonuspoint: info
+    org.apache.shenyu.test.lottery: debug
+    org.apache.shenyu.test: debug
+
